+++ packages/orq-rc/RELEASES.md
@@ -6419,3 +6419,13 @@ Based on:
 - [typescript v4.3.0-rc.26] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.26 - packages/orq-rc
+
+## 2026-02-05 10:02:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.704.0 (2.806.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.27] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.27 - packages/orq-rc